Fixed the mistake in the class being described
